fix: Remove redirect_facility to fix broken redirects (#678)
* fix: Remove redirect_facility to fix broken redirects

This came up in https://github.com/stackabletech/crddocs/issues/44.
Initial investigation revealed that the roles-and-role-groups.html
file was not present in the stable (24.7) and nightly (soon 24.11)
version of generated files when using the production antor playbook.

This playbook is used when Netlify deploys the docs using 'make
netlify-build'. Local builds using 'make' on the other hand included
the file in both versions and as such, the redirect works.

Finally, the 'redirect_facility: netlify' option was identified as
the cause. The generated '_redirects' file only contained a redirect
from concepts/roles-and-role-groups.html to concepts/stacklet. This
matched with the behaviour we experiences when surfing the live site.

Removing this option (and thus using the default facility) potentially
fixes the issue. Netlify already provides redirect using the "Pretty
URLs" post-processing feature.

References:

- https://docs.antora.org/antora/latest/playbook/urls-redirect-facility/#redirect-facility-key
- https://docs.netlify.com/site-deploys/post-processing/#post-processing-features
- https://docs.netlify.com/routing/redirects/#syntax-for-the-redirects-file

* feat: Add 'html_extension_style: indexify' to playbook